buffet: Make ExportedPropertySet threadsafe
In Chrome, we have both a DBus thread and some main thread (an 'origin'
thread).  Objects related to DBus functionality need to accomodate this
so that we can send this upstream.  This mostly just means that we need
to add assertions that developers are using the API correctly, so that
the use of weak pointers is safe.

BUG=chromium:360831
TEST=Unittests pass.  buffet_BasicDBusAPI still passes.

Change-Id: Ibb48a5e65c7cb02e5edce9cbf85432bed70d7686
Reviewed-on: https://chromium-review.googlesource.com/193505
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Tested-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>