libweave: Add 'reason' to backup command fetch requests
When XMPP channel is active we periodically (once every 30 minutes)
do a command fetch from the server. If XMPP channel doesn't miss any
commands, this fetch of command queue should never return any commands.

To help debug this on the server side, added "reason=just_in_case" to
such "backup" requests so we can identify if ever the server has to
return something for it. In addition, other reasons are added for
different fetch request origins:

- device_start
- regular_pull
- new_command
- just_in_case

BUG: 23321447
Change-Id: I6f2daf30424bc54a7b6d7d06052509e902c7ce1a
Reviewed-on: https://weave-review.googlesource.com/1391
Reviewed-by: Vitaly Buka <vitalybuka@google.com>