Remove hardcoded paths
The gyp files made some assumptions about where they were
in relation to each other, and that makes it tough to use
them as dependencies in a new project. This CL makes all
of the paths relative.

It also changes how the lib search path is defined. The
previous version placed a -L option in cflags. This
prevents the gyp generators from modifying the path,
so if you change to a different directory for output,
all of the third-party libs disappear. Fixed by specifying
the search path using gyp's library_dirs variable. Tested
with the ninja and cmake generators.

Change-Id: I489bfa33a97d28df0af77a699896be180ca9384b
Reviewed-on: https://weave-review.googlesource.com/1441
Reviewed-by: Vitaly Buka <vitalybuka@google.com>