Make sure example command definitions include 'minimalRole'
Server doesn't work correctly if minimalRole is not specified.
It just fails to forward a command to the device.

Also since Command::GetParameters returns a unique_ptr (rvalue),
it is generally a bad pattern to use cmd->GetParameter()->Get...
especially with dictionary and list values because by the time
the pointer to inner object is used the outer object will be
destroyed and access-after-free errors will occur.

Change-Id: I33fd3fadbf993ce172725d17befef016089fc0a7
Reviewed-on: https://weave-review.googlesource.com/1653
Reviewed-by: Vitaly Buka <vitalybuka@google.com>