Move deprecated APIs on Device object to the end of the interface
Move the old (obsolete) methods to the end of Device interface and
mark them with LIBWEAVE_DEPRECATED attribute. This attribute is
not fully enabled yet because these APIs are still used throughout
libweave and samples, so the warning will break the build.

However, once all this code is swept and all call sites are switched
to new component model, uncommend the definition of LIBWEAVE_DEPRECATED.

BUG: 25917706
Change-Id: I247bf74c9c94b22777e25d7021ff2c24e273c5e7
Reviewed-on: https://weave-review.googlesource.com/1781
Reviewed-by: Vitaly Buka <vitalybuka@google.com>