Add ClaimRootClientAuthToken and ConfirmRootClientAuthToken
Claim* call will create temporarily secret and set that as primary after
Confirm* is called.

Local client needs to be able to claim control of unclaimed device.
Device should be claimed no more than once, so if Claim was called and
client didn't not get response, we will have locked account. Confirm* is
used as that confirmation.

BUG=25766815

Change-Id: Id744f98788abe70a42b32c4a6d796e7ff74c3936
Reviewed-on: https://weave-review.googlesource.com/1947
Reviewed-by: Alex Vakulenko <avakulenko@google.com>