Remove internal/missing export includes
The behavior of LOCAL_WHOLE_STATIC_LIBRARIES is changing to also
reexport include directories in addition to every object in the library.
So remove the export of $(LOCAL_PATH) from the static libraries so that
the shared library exported include paths stay the same.

The ./external/ directory does not exist, so do not export it.

Change-Id: Iaff4ecc82a58c4b2b9b12f70c18fbc9523f9c4cb