libweave: Always fetch commands from server, even if XMPP delivers JSON
XMPP notification may include the command payload in its message. If it
it does, we normally don't go to the cloud server and fetch the current
command queue. However if XMPP notification is missed, there is no easy
way of obtaining the missed command.

If we use the notification as a cue to pull the current command queue
from the server, we would also pull any missed commands from earlier
missed notifications.

BUG: None
Change-Id: I85f0c2e58df7c2e6d2ce1be034d3c367870f7f7a
Reviewed-on: https://weave-review.googlesource.com/3170
Reviewed-by: Robert Ginda <rginda@google.com>