libweave: Update comment on handling new commands
The old comment in DeviceRegistrationInfo::OnCommandCreated() was left
over from when the code would use the command received via the
notification channel. Now the code always fetches commands from the
command queue when it receives a notification that there is a new
command. The code ignores the actual command in the notification
payload.

Change-Id: I1b7f7eedc337280035a5a6ec45f5935f780cac7f
Reviewed-on: https://weave-review.googlesource.com/3190
Reviewed-by: John Mccullough <jmccullough@google.com>